[fix] downloader

+++ bot/utils/downloader.py
@@ -1,20 +1,16 @@
-import asyncio
-import os
 import shutil
-import subprocess
+import os
 import uuid
+import asyncio
+from bot.utils.video_info import get_video_info
+from bot.utils.log import log_action
+from bot.database.mongo import save_to_cache, get_from_cache
+from bot.config import bot, semaphore
+import subprocess
 from aiogram.types import FSInputFile
 
-from bot.config import bot
-from bot.database.mongo import save_to_cache, get_from_cache
-from bot.utils.log import log_action
-from bot.utils.video_info import get_video_info  # Импортируем новый метод
-
 downloading_status = {}
-max_concurrent_downloads = 10  # Максимальное количество одновременных загрузок
-semaphore_downloads = asyncio.Semaphore(max_concurrent_downloads)  # Ограничение на количество одновременных загрузок
 
-# Асинхронная загрузка с использованием yt-dlp для получения ссылки
 async def download_media_async(url, download_type="video", quality="720", output_dir="downloads"):
     os.makedirs(output_dir, exist_ok=True)
     random_name = str(uuid.uuid4())
@@ -49,7 +45,6 @@ async def download_media_async(url, download_type="video", quality="720", output
         log_action("Ошибка скачивания", stderr.decode())
         return None
 
-# Асинхронная загрузка и отправка файлов
 async def download_and_send(user_id, url, download_type, quality):
     if downloading_status.get(user_id):
         await bot.send_message(user_id, "Видео уже скачивается. Пожалуйста, подождите.")
@@ -57,8 +52,7 @@ async def download_and_send(user_id, url, download_type, quality):
 
     downloading_status[user_id] = True
 
-    # Используем семафор для ограничений на количество одновременных задач
-    async with semaphore_downloads:
+    async with semaphore:
         video_id, title, _ = await get_video_info(url)
         if not video_id:
             await bot.send_message(user_id, "Не удалось извлечь информацию о видео.")
@@ -68,34 +62,30 @@ async def download_and_send(user_id, url, download_type, quality):
         cached_file_id = await get_from_cache(video_id, download_type, quality)
         if cached_file_id:
             if download_type == "video":
-                await bot.send_video(user_id, video=cached_file_id, caption=f"Ваше видео готово: {title}", supports_streaming=True)
+                await bot.send_video(user_id, video=cached_file_id, caption=f"Ваше видео готово: {title}")
             else:
                 await bot.send_audio(user_id, audio=cached_file_id, caption=f"Ваше аудио готово: {title}")
             downloading_status.pop(user_id, None)
             return
 
-        # Параллельная загрузка и отправка файла
-        async def download_and_send_file():
-            output_file = await download_media_async(url, download_type, quality)
-            if not output_file or not os.path.exists(output_file):
-                await bot.send_message(user_id, "Ошибка скачивания.")
-                log_action("Ошибка: файл отсутствует", f"Путь: {output_file}")
-                downloading_status.pop(user_id, None)
-                return
+        output_file = await download_media_async(url, download_type, quality)
+        if not output_file or not os.path.exists(output_file):
+            await bot.send_message(user_id, "Ошибка скачивания.")
+            log_action("Ошибка: файл отсутствует", f"Путь: {output_file}")
+            downloading_status.pop(user_id, None)
+            return
 
+        if output_file:
             file_to_send = FSInputFile(output_file)
             if download_type == "video":
-                message = await bot.send_video(user_id, video=file_to_send, caption=f"Ваше видео готово: {title}", supports_streaming=True)
+                message = await bot.send_video(user_id, video=file_to_send, caption=f"Ваше видео готово: {title}")
                 await save_to_cache(video_id, download_type, quality, message.video.file_id)
             else:
                 message = await bot.send_audio(user_id, audio=file_to_send, caption=f"Ваше аудио готово: {title}")
                 await save_to_cache(video_id, download_type, quality, message.audio.file_id)
             os.remove(output_file)
 
-            downloading_status.pop(user_id, None)
-
-        # Запускаем параллельно с ограничением количества задач
-        await download_and_send_file()
+        downloading_status.pop(user_id, None)
 
 def check_ffmpeg_installed():
     if not shutil.which("ffmpeg"):
